fix: hmr issues with non-react component exports
