+++ extensions/vscode/package.json
@@ -2,7 +2,7 @@
   "name": "continue",
   "icon": "media/icon.png",
   "author": "Continue Dev, Inc",
-  "version": "0.9.210",
+  "version": "0.9.211",
   "repository": {
     "type": "git",
     "url": "https://github.com/continuedev/continue"
