+++ gui/src/components/OnboardingCard/index.ts
@@ -1,4 +1,3 @@
 export * from "./OnboardingCard";
-export { default } from "./OnboardingCard";
 export * from "./utils";
 export * from "./hooks/useOnboardingCard";
