+++ gui/src/components/loaders/ProgressBar.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
-import { StyledTooltip, lightGray, vscForeground } from "..";
+import { lightGray, vscForeground } from "..";
 import AddModelForm from "../../forms/AddModelForm";
 import {
   setDialogMessage,
@@ -11,6 +11,7 @@ import {
 } from "../../redux/slices/uiStateSlice";
 import { getFontSize } from "../../util";
 import { FREE_TRIAL_LIMIT_REQUESTS } from "../../util/freeTrial";
+import { ToolTip } from "../gui/Tooltip";
 
 const ProgressBarWrapper = styled.div`
   width: 100px;
@@ -71,9 +72,9 @@ const ProgressBar = ({ completed, total }: ProgressBarProps) => {
         </div>
         {tooltipPortalDiv &&
           ReactDOM.createPortal(
-            <StyledTooltip id="usage_progress_bar" place="top">
+            <ToolTip id="usage_progress_bar" place="top">
               Configure a model above in order to continue
-            </StyledTooltip>,
+            </ToolTip>,
             tooltipPortalDiv,
           )}
       </>
@@ -117,9 +118,9 @@ const ProgressBar = ({ completed, total }: ProgressBarProps) => {
 
       {tooltipPortalDiv &&
         ReactDOM.createPortal(
-          <StyledTooltip id="usage_progress_bar" place="top">
+          <ToolTip id="usage_progress_bar" place="top">
             {`Click to use your own API key or local LLM (required after ${FREE_TRIAL_LIMIT_REQUESTS} inputs)`}
-          </StyledTooltip>,
+          </ToolTip>,
           tooltipPortalDiv,
         )}
     </>
