+++ gui/src/components/modelSelection/ModelCard.tsx
@@ -11,7 +11,8 @@ import { PackageDimension } from "../../pages/AddNewModel/configs/models";
 import { providers } from "../../pages/AddNewModel/configs/providers";
 import HeaderButtonWithText from "../HeaderButtonWithText";
 import InfoHover from "../InfoHover";
-import ModelProviderTag, { ModelProviderTags } from "./ModelProviderTag";
+import { ModelProviderTag } from "./ModelProviderTag";
+import { ModelProviderTags } from "./utils";
 
 interface ModelCardProps {
   title: string;
@@ -42,11 +43,11 @@ const Div = styled.div<{ color: string; disabled: boolean; hovered: boolean }>`
     opacity: 0.5;
     `
       : props.hovered
-        ? `
+      ? `
     border: 1px solid ${props.color};
     background-color: ${props.color}22;
     cursor: pointer;`
-        : ""}
+      : ""}
 `;
 
 const DimensionsDiv = styled.div`
@@ -147,7 +148,9 @@ function ModelCard(props: ModelCardProps) {
           <h3>{props.title}</h3>
         </div>
 
-        {props.tags?.map((tag, i) => <ModelProviderTag key={i} tag={tag} />)}
+        {props.tags?.map((tag, i) => (
+          <ModelProviderTag key={i} tag={tag} />
+        ))}
 
         <p>{props.description}</p>
