+++ gui/src/components/modelSelection/ModelSelectionListbox.tsx
@@ -111,7 +111,7 @@ function ModelSelectionListbox({
 }: ModelSelectionListboxProps) {
   return (
     <StyledListbox value={selectedProvider} onChange={setSelectedProvider}>
-      <div className="relative mt-1">
+      <div className="relative mt-1 mb-2">
         <StyledListboxButton>
           <span className="flex items-center">
             {window.vscMediaUrl && selectedProvider.icon && (
@@ -160,7 +160,7 @@ function ModelSelectionListbox({
                     <span className="text-md">{option.title}</span>
 
                     {selected ? (
-                      <span className="inset-y-0 ml-auto flex items-center pl-3 text-amber-600">
+                      <span className="inset-y-0 ml-auto flex items-center pl-3">
                         <CheckIcon className="h-5 w-5" aria-hidden="true" />
                       </span>
                     ) : null}
